Move ibrahim files to match current directory structure
